Archive/backend_old/requirements.txt
google-generativeai>=0.3.0
PyPDF2>=3.0.0
python-docx>=0.8.11
sentence-transformers>=2.2.0
faiss-cpu>=1.9.0
numpy>=1.24.0
pandas>=2.0.0
python-dotenv>=1.0.0
flask>=2.3.0
flask-cors>=4.0.0